+++ app/src/main/java/com/anytypeio/anytype/di/feature/settings/AboutAppDI.kt
@@ -1,11 +1,9 @@
 package com.anytypeio.anytype.di.feature.settings
 
 import com.anytypeio.anytype.core_utils.di.scope.PerScreen
-import com.anytypeio.anytype.domain.auth.interactor.GetCurrentAccount
+import com.anytypeio.anytype.domain.auth.interactor.GetAccount
 import com.anytypeio.anytype.domain.auth.interactor.GetLibraryVersion
 import com.anytypeio.anytype.domain.auth.repo.AuthRepository
-import com.anytypeio.anytype.domain.block.repo.BlockRepository
-import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.ui.settings.AboutAppFragment
 import com.anytypeio.anytype.ui_settings.about.AboutAppViewModel
 import dagger.Module
@@ -32,10 +30,10 @@ object AboutAppModule {
     @Provides
     @PerScreen
     fun provideViewModelFactory(
-        getCurrentAccount: GetCurrentAccount,
+        getAccount: GetAccount,
         getLibraryVersion: GetLibraryVersion
     ): AboutAppViewModel.Factory = AboutAppViewModel.Factory(
-        getCurrentAccount = getCurrentAccount,
+        getAccount = getAccount,
         getLibraryVersion = getLibraryVersion
     )
 
@@ -43,9 +41,8 @@ object AboutAppModule {
     @Provides
     @PerScreen
     fun provideGetAccountUseCase(
-        repo: BlockRepository,
-        builder: UrlBuilder
-    ): GetCurrentAccount = GetCurrentAccount(repo = repo, builder = builder)
+        repo: AuthRepository
+    ): GetAccount = GetAccount(repo = repo)
 
     @JvmStatic
     @Provides